+++ ec-prueba-unexus-servicio-tony-veas/src/test/java/ec/unexus/prueba/servicio/tony/veas/EcPruebaUnexusServicioTonyVeasApplicationTests.java
@@ -1,4 +1,4 @@
-package ec.unexus.pruebaserviciotonyveas;
+package ec.unexus.prueba.servicio.tony.veas;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
